Update references to django-admin in docs
Wherever we reference the django-admin command in the docs, update
it to reference `django` or “The django command”.

Rename the `django-admin` directive to `django-cmd`, and the 
`django-admin-option` directive to `django-cmd-option`. If we’re 
comfortable with it, we can drop the `-cmd`, but it seemed better to 
identify the places to change first.

Rename the `djadmin` role to `djcmd`. This could similarly be called 
`django` or `dj` if we preferred.

The instructions for generating the man page prefer the `django` 
spelling, even for the `django-admin` man page.